requirements.txt
torch
opencv-python
numpy
joblib
Pillow
torchvision
rembg
matplotlib
scikit-learn